requirements.txt
taipy==3.0.0
openai==1.3.7
pillow==10.2.0